Create ParentProblem.js
This is designed to demonstrate an apparent issue with adding an item positioned relative to its parent.  As far as I can determine, this works as advertised for Rect, Circle, Arrow, Crescent, ArcSegment, Cube, Gear, Isobox, Isotube, Oval, and Poly objects but it fails for Heart, BezierCurve, BezierSegment, Curve, CurveSegment, Grid, Line, and Path objects.  

In this file, the default properties of objects and the ignoring of inapplicable object properties (like "rotation" for Cube) are utilized.

This came about after heavily using the Arrow object, I was unable to simply replace it with the Heart object.